[#109] 미터 단위 표시 오류 및 키워드 공백 시 빈리스트 반환

+++ src/main/kotlin/com/deepromeet/atcha/location/infrastructure/client/response/TMapPOIResponse.kt
@@ -36,7 +36,7 @@ data class TMapPOIResponse(
                 Location(name, Coordinate(noorLat.toDouble(), noorLon.toDouble())),
                 getBusinessCategory(),
                 getLoadAddress(),
-                radius.toDouble().toInt()
+                radius.toDouble()
             )
         }
